@@ -1,8 +1,7 @@
 #! /usr/bin/env python
-import pkg_resources
+import importlib.metadata
 
-
-__version__ = pkg_resources.get_distribution("{{ cookiecutter.package_name }}").version
+__version__ = importlib.metadata.version("{{ cookiecutter.package_name }}")
 
 
 from .bmi import (
